# lab3/docs/swagger.yaml
basePath: /api/v1
definitions:
  serializer.BigORequestJSON:
    properties:
      bigo_request_id:
        type: integer
      calculated_complexity:
        type: string
      calculated_time:
        type: number
      creator_login:
        type: string
      date_create:
        type: string
      date_finish:
        type: string
      date_update:
        type: string
      moderator_login:
        type: string
      status:
        type: string
    type: object
...
  serializer.StatusJSON:
    properties:
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@bigorequest.com
    name: API Support
    url: http://localhost:8080
  description: API для управления расчётами времени и сложности Классов сложности
  license:
    name: MIT
  title: Big O Request API
  version: "1.0"
paths:
  /bigorequest/{id}:
    get:
      description: Возвращает полную информацию о заявке
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Данные заявки с классами сложности
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Заявка не найдено
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить заявку по ID
      tags:
      - bigorequests
  /bigorequest/{id}/delete-bigorequest:
    delete:
      description: Выполняет логическое удаление заявки
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Статус удаления
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Заявка не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить заявку
      tags:
      - bigorequests
  /bigorequest/{id}/edit-bigorequest:
    put:
      consumes:
      - application/json
      description: Обновляет данные заявки
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: Новые данные заявки
        in: body
        name: bigorequest
        required: true
        schema:
          $ref: '#/definitions/serializer.BigORequestJSON'
      produces:
      - application/json
      responses:
        "200":
          description: Обновленная заявка
          schema:
            $ref: '#/definitions/serializer.BigORequestJSON'
        "400":
          description: Неверные данные
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Заявка не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Изменить заявку
      tags:
      - bigorequests
  /bigorequest/{id}/finish-bigorequest:
    put:
      consumes:
      - application/json
      description: Изменяет статус заявки (только для модераторов)
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: Новый статус
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/serializer.StatusJSON'
      produces:
      - application/json
      responses:
        "200":
          description: Результат модерации
          schema:
            $ref: '#/definitions/serializer.BigORequestJSON'
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Заявка не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Завершить заявку
      tags:
      - bigorequests
  /bigorequest/{id}/form-bigorequest:
    put:
      description: Переводит заявку в статус "formed"
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Сформированная заявка
          schema:
            $ref: '#/definitions/serializer.BigORequestJSON'
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Заявка не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Сформировать заявку
      tags:
      - bigorequests
  /bigorequest/all-bigo_requests:
    get:
      description: Возвращает заявки с возможностью фильтрации по датам и статусу
      parameters:
      - description: Начальная дата (YYYY-MM-DD)
        in: query
        name: from-date
        type: string
      - description: Конечная дата (YYYY-MM-DD)
        in: query
        name: to-date
        type: string
      - description: Статус заявки
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список заявок
          schema:
            items:
              $ref: '#/definitions/serializer.BigORequestJSON'
            type: array
        "400":
          description: Неверный формат даты
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить список заявок на расчёт
      tags:
      - bigorequests
  /bigorequest/bigorequest-cart:
    get:
      description: Возвращает информацию о текущей заявке-черновике на расчёт пользователя
      produces:
      - application/json
      responses:
        "200":
          description: Данные корзины заявки-черновика
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить корзину расчёта
      tags:
      - bigorequests
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"